feat: separated bunch of schemas into individual logical schemas, adding validations for inputs
